Adding one time cookie. When we close browser, the cookie destroies itself

+++ signout.php
@@ -1,7 +1,18 @@
 <?php
-    $cser=mysqli_connect("localhost","root","","JOB_BOARD") or die("connection failed:".mysqli_error());
-    $del1 = "DELETE FROM cookies where hash_id = '".$_COOKIE["login"]."'";
-    setcookie("login","",time()-1);//for delete the cookie //destroy the cookie 
-    $cser->query($del1);
+    include "db_connection.php";
+
+    $cser=OpenCon() or die("connection failed:".mysqli_error());
+
+    if(isset($_COOKIE["login_one_time"]) ){ 
+        $del1 = "DELETE FROM cookies where hash_id = '".$_COOKIE["login_one time"]."'";
+        setcookie("login_one_time","",time()-1); 
+        $cser->query($del1);
+    }
+    if(isset($_COOKIE["login"])){
+        $del1 = "DELETE FROM cookies where hash_id = '".$_COOKIE["login"]."'";
+        setcookie("login","",time()-1);
+        $cser->query($del1);
+    }
+
     header("location:index.php"); 
 ?>
